Move ClusterSettingsPage and DeregistrationModal to pages
